#163: use deleteReply to consolidate viewgearsales flow - fix gear potentially selling for fractional gold
